Split into parser submodule
